requirements.txt
fastapi
uvicorn
pandas
requests
openai
streamlit
python-dotenv